chore: apply more complicated lints (#2576)
<!--
Please provide as much information as possible about what your PR aims
to do.
PRs with no description will most likely be closed until more
information is provided.
If you're planing on changing fundamental behaviour or add big new
features, please open a GitHub Issue first before starting to work on
it.
If it's not something big and you still want to contact us about it,
feel free to do so !
-->

### Problem description
<!-- Describe the bug that you fixed/feature request that you
implemented, or link to an existing issue describing it -->

### Implementation description
<!-- Explain what you did to correct the problem -->

### Screenshots
<!-- If your change is visual, take a screenshot showing it. Ideally,
make before/after sceenshots -->

### Additional things
<!-- Anything else you would like to say -->

+++ main/gui/source/init/splash_window.cpp
@@ -589,8 +589,8 @@ namespace hex::init {
         for (auto &highlight : m_highlights) {
             u32 newPos = lastPos + lastCount + (rng() % 35);
             u32 newCount = (rng() % 7) + 3;
-            highlight.start.x = float(newPos % 13);
-            highlight.start.y = float(newPos / 13);
+            highlight.start.x = newPos % 13;
+            highlight.start.y = int(newPos / 13);
             highlight.count = newCount;
 
             highlight.color = getHighlightColor(index);
